+++ packages/apidom-ls/src/config/asyncapi/bindings/ws/channel-binding/lint/0-1-0/index.ts
