+++ packages/apidom-ls/src/config/asyncapi/bindings/ws/channel-binding/lint/index.ts
@@ -1,4 +1,4 @@
-import channelBinding0_1_0Lints from './0.1.0';
+import channelBinding0_1_0Lints from './0-1-0';
 import channelBindingLatestLints from './latest';
 import bindingVersionTypeLint from './binding-version--type';
